// src/app/paginas/interfaces/model.productos.ts
export interface ModuleProductos {
    product_id:       number;
    product_title:    string;
    product_description: string;
    product_cost:     number;
    product_price:     number;
    product_stock:    number;
    product_type:    string;
    product_image:    string;
    product_state:    boolean;
}